DNN-8667: disable purge log buffer by default when event viewer control loaded.

+++ DNN Platform/Admin Modules/Dnn.Modules.LogViewer/LogViewer.ascx.cs
@@ -367,7 +367,12 @@ namespace Dnn.Modules.LogViewer
                 {
                     txtSubject.Text = Localization.GetString("LogEntryDefaultSubject", LocalResourceFile);
                     txtMessage.Text = Localization.GetString("LogEntryDefaultMsg", LocalResourceFile);
-                    LogController.Instance.PurgeLogBuffer();
+
+                    if (AutoPurgeLogBuffer)
+                    {
+                        LogController.Instance.PurgeLogBuffer();
+                    }
+
                     if (Request.QueryString["PageRecords"] != null)
                     {
                         ddlRecordsPerPage.SelectedValue = Request.QueryString["PageRecords"];
@@ -497,6 +502,14 @@ namespace Dnn.Modules.LogViewer
         }
 
         public string EventFilter { get; set; }
+
+        public bool AutoPurgeLogBuffer
+        {
+            get
+            {
+                return Settings.ContainsKey("AutoPurgeLogBuffer")  && Convert.ToBoolean(Settings["AutoPurgeLogBuffer"]);
+            }
+        }
 
         #endregion
